Start a week-log blog.
Instead of starting afresh, which was resisting my efforts,
just turn learn-generating-functions blog into a template and use it.

The new repo seems to work fine as a blog, so now I'm fixing the title
and adding a couple of new posts. If that works, I'll get rid of the old
posts and be good to go!